Rename package to googleclouddns
